completed cloud functions
