+++ functions/package.json
@@ -14,11 +14,14 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "@google-cloud/logging": "^8.0.0",
     "@google-cloud/storage": "^5.1.1",
+    "child-process-promise": "^2.2.1",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.7.0",
+    "mkdirp-promise": "^5.0.1",
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
